Minor Changes
Wrapped `stb_image` into a module. Tried the same thing for the other libraries but didn't success.

Removed `XrSpaceTransform`.

+++ NaiveOpenXrGame/sources/Renderer/SwapChain.cpp
@@ -3,7 +3,7 @@
 
 Noxg::SwapChain::SwapChain(vk::Device device, vk::RenderPass renderPass, std::vector<vk::Image>& images, vk::Format format, vk::Rect2D rect)
 {
-	this->length = images.size();
+	this->length = static_cast<uint32_t>(images.size());
 	this->format = format;
 	this->rect = rect;
 	this->device = device;
